+++ SonsSdk/DebugTools.cs
@@ -1,7 +1,10 @@
 ﻿using System.ComponentModel;
+using System.Reflection;
+using HarmonyLib;
 using Il2CppInterop.Runtime.Injection;
 using Il2CppInterop.Runtime.InteropTypes.Fields;
 using RedLoader;
+using Shapes;
 using Sons.Items.Core;
 using Sons.Weapon;
 using TheForest.Utils;
@@ -15,6 +18,9 @@ public class DebugTools
 {
     private static Material _glMaterial = GetHdrpMaterial(Color.red);
     
+    private static MethodBase _ueInspectMethod;
+    private static PropertyInfo _ueShowMenuProperty;
+
     public static Material GetHdrpMaterial(Color? color = null)
     {
         var material = new Material(Shader.Find("HDRP/Lit"));
@@ -24,35 +30,6 @@ public class DebugTools
         }
         
         return material;
-    }
-
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public static void DrawGlLine(Vector3 start, Vector3 end, float thickness = 1)
-    {
-        _glMaterial.SetPass(0);
-        GL.PushMatrix();
-        
-        Vector3 direction = (end - start).normalized;
-        
-        Vector3 normal = Vector3.Cross(direction, Camera.main.transform.forward).normalized;
-        
-        float halfThickness = thickness * 0.5f;
-        
-        Vector3 v1 = start - normal * halfThickness;
-        Vector3 v2 = start + normal * halfThickness;
-        Vector3 v3 = end + normal * halfThickness;
-        Vector3 v4 = end - normal * halfThickness;
-        
-        GL.Begin(0x0007);
-        GL.Color(Color.red);
-        
-        GL.Vertex(v1);
-        GL.Vertex(v2);
-        GL.Vertex(v3);
-        GL.Vertex(v4);
-        
-        GL.End();
-        GL.PopMatrix();
     }
 
     public static GameObject CreatePrimitive(PrimitiveType type, Vector3? pos = null, Color? color = null)
@@ -66,6 +43,32 @@ public class DebugTools
         }
 
         return go;
+    }
+
+    public static Disc CreateDisc(Vector3 position, float radius = 0.2f, Color? color = null, bool alwaysInFront = true, bool billboard = true)
+    {
+        var disc = new GameObject("Disc").AddComponent<Disc>();
+        disc.transform.position = position;
+        disc.Radius = radius;
+        disc.Color = color ?? Color.red;
+        if(alwaysInFront)
+            disc.ZTest = CompareFunction.Always;
+        if(billboard)
+            disc.Geometry = DiscGeometry.Billboard;
+
+        return disc;
+    }
+    
+    public static Cuboid CreateCuboid(Vector3 position, Vector3? size = null, Color? color = null, bool alwaysInFront = true)
+    {
+        var cuboid = new GameObject("Cuboid").AddComponent<Cuboid>();
+        cuboid.transform.position = position;
+        cuboid.Size = size ?? new(0.5f, 0.5f, 0.5f);
+        cuboid.Color = color ?? Color.red;
+        if(alwaysInFront)
+            cuboid.ZTest = CompareFunction.Always;
+
+        return cuboid;
     }
 
     internal static LineRenderer GetNewLineRenderer(bool useWorldSpace = true, float width = 0.2f)
@@ -83,6 +86,51 @@ public class DebugTools
         lineRenderer.gameObject.SetActive(true);
         
         return lineRenderer;
+    }
+
+    /// <summary>
+    /// Will inspect an object in Unity Explorer if installed.
+    /// </summary>
+    /// <param name="obj"></param>
+    /// <param name="showExporer">if true opens Unity Explorer too</param>
+    public static void Inspect(object obj, bool showExporer = false)
+    {
+        if (SonsMod.RegisteredMods.All(x => x.ID != "UnityExplorer"))
+        {
+            RLog.Msg(System.Drawing.Color.IndianRed, "Unity Explorer is not installed!");
+            return;
+        }
+
+        if (_ueInspectMethod == null)
+        {
+            _ueInspectMethod = AccessTools.TypeByName("UnityExplorer.InspectorManager").GetMethods(BindingFlags.Public | BindingFlags.Static).FirstOrDefault(x =>
+            {
+                return x.Name == "Inspect" && x.GetParameters().First().ParameterType == typeof(object);
+            });
+        }
+        
+        if (_ueInspectMethod == null)
+        {
+            RLog.Msg(System.Drawing.Color.IndianRed, "Couldn't get the inspect method for Unity Explorer!");
+            return;
+        }
+        
+        _ueInspectMethod.Invoke(null, new[] { obj, null });
+
+        if (!showExporer)
+            return;
+        
+        if (_ueShowMenuProperty == null)
+        {
+            _ueShowMenuProperty = AccessTools.TypeByName("UnityExplorer.UI.UIManager").GetProperty("ShowMenu");
+        }
+        
+        if (_ueShowMenuProperty == null)
+        {
+            return;
+        }
+        
+        _ueShowMenuProperty.SetValue(null, true);
     }
     
     /// <summary>
@@ -146,52 +194,5 @@ public class DebugTools
             Object.Destroy(Transform.gameObject);
             _isDestroyed = true;
         }
-    }
-}
-
-[EditorBrowsable(EditorBrowsableState.Never)]
-public class ImLineDrawer
-{
-    public Vector3 Start;
-    public Vector3 End;
-    
-    private bool _isRendering;
-    
-    public static ImLineDrawer StartNew()
-    {
-        var drawer = new ImLineDrawer();
-        drawer.BeginRendering();
-        return drawer;
-    }
-
-    public void SetLine(Vector3 start, Vector3 end)
-    {
-        Start = start;
-        End = end;
-    }
-
-    public void BeginRendering()
-    {
-        if (_isRendering)
-            return;
-        
-        _isRendering = true;
-        
-        SdkEvents.OnCameraRender.Subscribe(OnRender);
-    }
-    
-    public void EndRendering()
-    {
-        if (!_isRendering)
-            return;
-        
-        _isRendering = false;
-        
-        SdkEvents.OnCameraRender.Unsubscribe(OnRender);
-    }
-    
-    private void OnRender(ScriptableRenderContext context, Il2CppSystem.Collections.Generic.List<Camera> cams)
-    {
-        DebugTools.DrawGlLine(Start, End);
     }
 }