+++ SonsSdk/Savegame/RedloaderModSerializer.cs
@@ -1,9 +1,11 @@
+using System.Diagnostics;
 using System.Drawing;
 using Il2CppInterop.Runtime.Injection;
 using Il2CppSystem.Collections.Generic;
 using Newtonsoft.Json;
 using RedLoader;
 using RedLoader.TinyJSON;
+using RedLoader.Utils;
 using Sons.Save;
 using Object = Il2CppSystem.Object;
 using CppString = Il2CppSystem.String;
@@ -45,21 +47,38 @@ public class RedloaderModSerializer : Object
     
     ModSaveData OnSerialize()
     {
-        RLog.Msg(Color.Orange, new string('=', 50));
-        RLog.Msg(Color.Orange, "SERIALIZING MOD DATA");
-        RLog.Msg(Color.Orange, new string('=', 50));
+        RLog.Msg(Color.Yellow, new string('=', 50));
+        RLog.Msg(Color.Yellow, "SERIALIZING MOD DATA");
+
+        var sw = new Stopwatch();
 
         var datas = new System.Collections.Generic.List<NamedSaveData>();
         
         foreach (var (name, saveable) in SonsSaveTools.CustomSaveables)
         {
-            var namedSaveData = new NamedSaveData();
-            namedSaveData.Name = name;
-            namedSaveData.Data = saveable.Serialize();
-            datas.Add(namedSaveData);
+            try
+            {
+                var namedSaveData = new NamedSaveData();
+                namedSaveData.Name = name;
+                namedSaveData.Data = saveable.Serialize();
+                datas.Add(namedSaveData);
+                RLog.Msg(Color.Orange, $"Serializing {name} took {sw.ElapsedMilliseconds}ms");
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+
+            sw.Reset();
         }
 
-        return new ModSaveData(JSON.Dump(datas, EncodeOptions.NoTypeHints));
+        var modSaveData = new ModSaveData(JSON.Dump(datas, EncodeOptions.NoTypeHints));
+        sw.Stop();
+        RLog.Msg(Color.Orange, $"Finishing up took {sw.ElapsedMilliseconds}ms");
+        
+        RLog.Msg(Color.Yellow, new string('=', 50));
+
+        return modSaveData;
     }
     
     private ModSaveData DeserializeOverride(string data)
@@ -71,9 +90,8 @@ public class RedloaderModSerializer : Object
     // input data will be null here, so disregard it
     void OnDeserialize(ModSaveData _)
     {
-        RLog.Msg(Color.Orange, new string('=', 50));
-        RLog.Msg(Color.Orange, "DESERIALIZING MOD DATA");
-        RLog.Msg(Color.Orange, new string('=', 50));
+        RLog.Msg(Color.Yellow, new string('=', 50));
+        RLog.Msg(Color.Yellow, "DESERIALIZING MOD DATA");
 
         if (string.IsNullOrEmpty(_lastData))
         {
@@ -81,21 +99,53 @@ public class RedloaderModSerializer : Object
             return;
         }
 
+        var sw = new Stopwatch();
+
         var data = SaveGameManager.Deserialize<ModSaveData>(_lastData);
         _lastData = null;
 
-        var datas = JSON.Load(data.ModData.Get()).Make<System.Collections.Generic.List<NamedSaveData>>();
+        RLog.Msg(Color.Orange, $"SaveGameManager deserializing took {sw.ElapsedMilliseconds}ms");
+        sw.Reset();
+
+        System.Collections.Generic.List<NamedSaveData> datas;
         
+        try
+        {
+            datas = JSON.Load(data.ModData.Get()).Make<System.Collections.Generic.List<NamedSaveData>>();
+        }
+        catch (Exception e)
+        {
+            RLog.Error($"Error in deserializing mod data: {e}");
+            throw;
+        }
+
+        RLog.Msg(Color.Orange, $"Constructing mod list took {sw.ElapsedMilliseconds}ms");
+        sw.Reset();
+
         foreach (var modData in datas)
         {
             if (SonsSaveTools.CustomSaveables.TryGetValue(modData.Name, out var saveable))
             {
-                saveable.Deserialize(modData.Data);
+                try
+                {
+                    saveable.Deserialize(modData.Data);
+                    RLog.Msg(Color.Orange, $"Deserializing {modData.Name} took {sw.ElapsedMilliseconds}ms");
+                }
+                catch (Exception e)
+                {
+                    RLog.Error($"Error deserializing {modData.Name}: {e}");
+                    throw;
+                }
+                
+                sw.Reset();
                 continue;
             }
             
             RLog.Error($"Could not find saveable with name {modData.Name}");
         }
+        
+        sw.Stop();
+        RLog.Msg(Color.Yellow, new string('=', 50));
     }
 
     public RedloaderModSerializer(IntPtr ptr) : base(ptr)